[YS-531] refactor: UserDetailInfo 컴포넌트 분리
