+++ src/app/user/profile/mobile/components/MobileUserInfoSection/MobileUserInfoSection.css.ts
@@ -32,54 +32,3 @@ export const userBadge = style({
   display: 'inline-flex',
   alignItems: 'center',
 });
-
-export const userDetails = style({
-  ...fonts.body.small.R15,
-  display: 'flex',
-  flexDirection: 'column',
-});
-
-export const userDetailWrapper = style({
-  display: 'flex',
-  alignItems: 'center',
-  gap: '0.8rem',
-});
-
-export const verticalLine = style({
-  position: 'relative',
-
-  selectors: {
-    '&::after': {
-      content: '""',
-      width: '0.1rem',
-      height: '1.2rem',
-      position: 'absolute',
-      top: '50%',
-      right: 0,
-      transform: 'translateY(-50%)',
-      backgroundColor: colors.line03,
-    },
-  },
-});
-
-export const gender = style({
-  color: colors.text02,
-});
-
-export const birthDate = style({
-  color: colors.text02,
-});
-
-export const userID = style({
-  color: colors.text02,
-});
-
-export const userIDLabel = style({
-  color: colors.text03,
-});
-
-export const userIDWrapper = style({
-  display: 'flex',
-  alignItems: 'center',
-  gap: '0.8rem',
-});
