+++ src/app/user/profile/mobile/components/MobileUserInfoSection/MobileUserInfoSection.tsx
@@ -1,34 +1,11 @@
 'use client';
 
-import {
-  userInfoHeader,
-  userInfoContent,
-  userName,
-  userBadge,
-  userDetails,
-  verticalLine,
-  gender,
-  birthDate,
-  userIDLabel,
-  userDetailWrapper,
-  userIDWrapper,
-  userID,
-} from './MobileUserInfoSection.css';
+import { userInfoHeader, userInfoContent, userName, userBadge } from './MobileUserInfoSection.css';
+import UserDetailInfo from './UserDetailInfo/UserDetailInfo';
 
 import useUserInfo from '@/app/home/hooks/useUserInfo';
 import { colors } from '@/styles/colors';
 import { isParticipantInfo } from '@/utils/typeGuard';
-
-export const formatBirthDate = (date: string) => {
-  if (!date) return '';
-  return date.replace(/-/g, '. ') + '.';
-};
-
-const GENDER_LABEL = {
-  MALE: '남성',
-  FEMALE: '여성',
-  ALL: '전체',
-} as const;
 
 const MobileUserInfoSection = () => {
   const { userInfo } = useUserInfo();
@@ -51,19 +28,8 @@ const MobileUserInfoSection = () => {
             {isParticipant ? '참여자' : '연구자'}
           </span>
         </div>
-        <div className={userDetails}>
-          {isParticipant && (
-            <div className={userDetailWrapper}>
-              <span className={gender}>{GENDER_LABEL[userInfo.gender]}</span>
-              <span className={verticalLine} />
-              <span className={birthDate}>{formatBirthDate(userInfo.birthDate)}</span>
-            </div>
-          )}
-          <div className={userIDWrapper}>
-            <span className={userIDLabel}>ID</span>
-            <span className={userID}>{memberInfo?.oauthEmail}</span>
-          </div>
-        </div>
+
+        <UserDetailInfo userInfo={userInfo} />
       </div>
     </div>
   );
